hv: avoid using of mixed mode arithmetic
Avoid using of mixed mode arithmetic by using explicit casts

Tracked-On: #861

+++ hypervisor/arch/x86/cpu.c
@@ -636,7 +636,7 @@ void start_cpus(void)
 	/* Wait until global count is equal to expected CPU up count or
 	 * configured time-out has expired
 	 */
-	timeout = CONFIG_CPU_UP_TIMEOUT * 1000U;
+	timeout = (uint32_t)CONFIG_CPU_UP_TIMEOUT * 1000U;
 	while ((atomic_load16(&up_count) != expected_up) && (timeout != 0U)) {
 		/* Delay 10us */
 		udelay(10U);
@@ -664,7 +664,7 @@ void stop_cpus(void)
 	uint16_t pcpu_id, expected_up;
 	uint32_t timeout;
 
-	timeout = CONFIG_CPU_UP_TIMEOUT * 1000U;
+	timeout = (uint32_t)CONFIG_CPU_UP_TIMEOUT * 1000U;
 	for (pcpu_id = 0U; pcpu_id < phys_cpu_num; pcpu_id++) {
 		if (get_cpu_id() == pcpu_id) {	/* avoid offline itself */
 			continue;
